refactor: submission filter
