+++ wondrous-app/components/Common/TaskSubmission/submission.tsx
@@ -1,14 +1,13 @@
 import { useMutation } from '@apollo/client';
-import { CircularProgress, ClickAwayListener } from '@mui/material';
+import { CircularProgress } from '@mui/material';
 import { useMe } from 'components/Auth/withAuth';
 import { UPDATE_TASK_STATUS } from 'graphql/mutations';
-import { isEmpty, values } from 'lodash';
+import { isEmpty } from 'lodash';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { delQuery } from 'utils';
 import {
   ENTITIES_TYPES,
-  PAYMENT_STATUS,
   TASK_STATUS_DONE,
   TASK_STATUS_IN_PROGRESS,
   TASK_STATUS_IN_REVIEW,
@@ -21,16 +20,10 @@ import {
   SubmissionButtonTextHelper,
   SubmissionButtonWrapperBackground,
   SubmissionButtonWrapperGradient,
-  SubmissionFilterButtonIcon,
-  SubmissionFilterSelectButton,
-  SubmissionFilterSelectItem,
-  SubmissionFilterSelectMenu,
-  SubmissionFilterSelectPopper,
-  SubmissionFilterSelectRender,
-  SubmissionFilterStatusIcon,
   TaskSubmissionItemsWrapper,
   TaskSubmissionsFormInactiveWrapper,
 } from './styles';
+import { TaskSubmissionsFilter } from './submissionFilter';
 import { TaskSubmissionForm } from './submissionForm';
 import { SubmissionItem } from './submissionItem';
 import { SubmissionPayment } from './submissionPayment';
@@ -173,70 +166,6 @@ const TaskSubmissionList = ({
         );
       })}
     </TaskSubmissionItemsWrapper>
-  );
-};
-
-const TaskSubmissionsFilterSelected = ({ value }) => {
-  const text = isEmpty(value) ? 'Status' : value;
-  return (
-    <SubmissionFilterSelectRender>
-      <SubmissionFilterStatusIcon /> {text}
-    </SubmissionFilterSelectRender>
-  );
-};
-
-const filterOptions = {
-  awaitingReview: 'Awaiting Review',
-  changesRequested: 'Changes Requested',
-  approved: 'Approved',
-  approvedAndProcessingPayment: 'Approved And Processing Payment',
-  approvedAndPaid: 'Approved and Paid',
-};
-
-const isSubmissionStatus = ({ submission, value }) => {
-  const { awaitingReview, changesRequested, approved, approvedAndProcessingPayment, approvedAndPaid } = filterOptions;
-  const conditions = {
-    [awaitingReview]: !submission?.approvedAt && !submission?.changeRequestedAt && !submission.rejectedAt,
-    [changesRequested]: submission?.changeRequestedAt || submission?.rejectedAt,
-    [approved]: submission?.approvedAt,
-    [approvedAndProcessingPayment]: submission?.approvedAt && submission?.paymentStatus === PAYMENT_STATUS.PROCESSING,
-    [approvedAndPaid]: submission?.approvedAt && submission?.paymentStatus === PAYMENT_STATUS.PAID,
-  };
-  return conditions[value];
-};
-
-const TaskSubmissionsFilter = ({ fetchedTaskSubmissions, setFilteredSubmissions }) => {
-  const [selected, setSelected] = useState('');
-  const [anchorEl, setAnchorEl] = useState(null);
-  const handleClick = (event) => setAnchorEl(anchorEl ? null : event.currentTarget);
-  const handleClosePopper = () => setAnchorEl(null);
-  const open = Boolean(anchorEl);
-  const handleOnClick = (e) => {
-    const value = e.target.getAttribute('value');
-    setSelected(value);
-    const handleFilterStatus = (submission) => isSubmissionStatus({ submission, value });
-    setFilteredSubmissions(fetchedTaskSubmissions.filter(handleFilterStatus));
-    handleClosePopper();
-  };
-  if (isEmpty(fetchedTaskSubmissions)) return null;
-  return (
-    <ClickAwayListener onClickAway={handleClosePopper}>
-      <div>
-        <SubmissionFilterSelectButton onClick={handleClick} open={open}>
-          <TaskSubmissionsFilterSelected value={selected} />
-          <SubmissionFilterButtonIcon open={open} />
-        </SubmissionFilterSelectButton>
-        <SubmissionFilterSelectPopper open={open} anchorEl={anchorEl} placement="bottom-start" disablePortal={true}>
-          <SubmissionFilterSelectMenu>
-            {values(filterOptions).map((i) => (
-              <SubmissionFilterSelectItem key={i} value={i} onClick={handleOnClick}>
-                <SubmissionFilterStatusIcon /> {i}
-              </SubmissionFilterSelectItem>
-            ))}
-          </SubmissionFilterSelectMenu>
-        </SubmissionFilterSelectPopper>
-      </div>
-    </ClickAwayListener>
   );
 };
